Add GQL queries to components.
