+++ native/gql.ts
@@ -1,9 +1,6 @@
-import { GraphQLClient } from 'graphql-request'
+import ApolloClient from 'apollo-boost'
 
-import { getSdk } from './graphqlSdk'
-
-export const client = getSdk(
-  new GraphQLClient('https://bear-with-life-beta-api.herokuapp.com/v1/graphql', {
-    headers: { 'x-hasura-admin-secret': 'VacuumlabsIsTheBest' },
-  }),
-)
+export const client = new ApolloClient({
+  uri: 'https://bear-with-life-beta-api.herokuapp.com/v1/graphql',
+  headers: { 'x-hasura-admin-secret': 'VacuumlabsIsTheBest' },
+})
