+++ native/package.json
@@ -9,19 +9,21 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@apollo/react-components": "^3.1.4",
+    "@apollo/react-hooks": "^3.1.4",
     "@eva-design/eva": "^1.4.0",
-    "@graphql-codegen/typescript-graphql-request": "1.13.0",
+    "@graphql-codegen/typescript-react-apollo": "^1.13.1",
     "@react-native-community/masked-view": "0.1.5",
     "@react-navigation/native": "^5.1.4",
     "@react-navigation/stack": "^5.2.9",
     "@types/graphql": "^14.5.0",
     "@ui-kitten/components": "^4.4.1",
+    "apollo-boost": "^0.4.7",
     "appcenter": "^3.0.0",
     "appcenter-analytics": "^3.0.0",
     "appcenter-crashes": "^3.0.0",
     "expo": "~36.0.0",
     "graphql": "^15.0.0",
-    "graphql-request": "^1.8.2",
     "graphql-tag": "^2.10.3",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
