+++ native/src/components/screens/homeScreen.tsx
@@ -1,3 +1,4 @@
+import _ from 'lodash'
 import React from 'react'
 import { Dimensions, ScrollView } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
@@ -5,46 +6,33 @@ import { SafeAreaView } from 'react-native-safe-area-context'
 import { ChallengeCard } from '@components/ui/ChallengeCard'
 import { Layout, Text } from '@ui-kitten/components'
 
+import { useChallengesQuery } from '../../../graphqlSdk'
+
 const COLORS = {
   orange: '#ffc342',
   lightbrown: '#f5efe3',
   navy: '#0c2945',
   cyan: '#37a2a4',
-  white: '#ffffff',
 }
 
-const CHALLENGES = [
-  {
-    id: '1',
-    name: 'Call a family member',
-    color: COLORS.orange,
-    people: ['https://i.pravatar.cc/300', 'https://i.pravatar.cc/300', 'https://i.pravatar.cc/300'],
-  },
-  {
-    id: '2',
-    name: 'Tidy up around you',
-    color: COLORS.navy,
-    people: ['https://i.pravatar.cc/300', 'https://i.pravatar.cc/300'],
-  },
-  {
-    id: '3',
-    name: 'Call a family member',
-    color: COLORS.cyan,
-    people: ['https://i.pravatar.cc/300'],
-  },
-]
-
-const CHALLENGES2 = CHALLENGES.map(challenge => {
-  const { people, ...rest } = challenge
-  return { ...rest, color: COLORS.lightbrown }
-})
-
-const CHALLENGES3 = CHALLENGES.map(challenge => {
-  const { people, ...rest } = challenge
-  return rest
-})
-
 export const HomeScreen: React.FC = () => {
+  const { data, loading, error } = useChallengesQuery()
+
+  if (loading) {
+    return <Text>Loading...</Text>
+  }
+
+  if (error || !data) {
+    return <Text>Error!</Text>
+  }
+
+  const { challenge: challenges } = data
+
+  const challengesWithColors = challenges.map(challenge => ({
+    ...challenge,
+    color: _.sample(COLORS),
+  }))
+
   return (
     <Layout style={{ flex: 1 }} level="1">
       <Layout level="3">
@@ -57,7 +45,7 @@ export const HomeScreen: React.FC = () => {
             showsHorizontalScrollIndicator={false}
             contentContainerStyle={{ paddingHorizontal: 8 }}
           >
-            {CHALLENGES.map(challenge => (
+            {challengesWithColors.map(challenge => (
               <ChallengeCard
                 challenge={challenge}
                 key={challenge.id}
@@ -77,7 +65,7 @@ export const HomeScreen: React.FC = () => {
             showsHorizontalScrollIndicator={false}
             contentContainerStyle={{ paddingHorizontal: 8 }}
           >
-            {CHALLENGES2.map(challenge => (
+            {challengesWithColors.map(challenge => (
               <ChallengeCard
                 challenge={challenge}
                 key={challenge.id}
@@ -94,7 +82,7 @@ export const HomeScreen: React.FC = () => {
             showsHorizontalScrollIndicator={false}
             contentContainerStyle={{ paddingHorizontal: 8 }}
           >
-            {CHALLENGES3.map(challenge => (
+            {challengesWithColors.map(challenge => (
               <ChallengeCard
                 challenge={challenge}
                 key={challenge.id}
